Bundle Handling: Ensure proper return values for error codes

+++ core/net/uDTN/bundle.c
@@ -32,7 +32,7 @@
  * "Internal" functions
  */
 static uint8_t bundle_decode_block(struct mmem *bundlemem, uint8_t *buffer, int max_len);
-static uint8_t bundle_encode_block(struct bundle_block_t *block, uint8_t *buffer, uint8_t max_len);
+static int bundle_encode_block(struct bundle_block_t *block, uint8_t *buffer, uint8_t max_len);
 
 
 struct mmem * bundle_create_bundle()
@@ -64,7 +64,7 @@ struct mmem * bundle_create_bundle()
 	return &bs->bundle;
 }
 
-uint8_t bundle_add_block(struct mmem *bundlemem, uint8_t type, uint8_t flags, uint8_t *data, uint8_t d_len)
+int bundle_add_block(struct mmem *bundlemem, uint8_t type, uint8_t flags, uint8_t *data, uint8_t d_len)
 {
 	struct bundle_t *bundle;
 	struct bundle_block_t *block;
@@ -382,7 +382,7 @@ static uint8_t bundle_decode_block(struct mmem *bundlemem, uint8_t *buffer, int 
 	return offs + block->block_size;
 }
 
-uint8_t bundle_encode_bundle(struct mmem *bundlemem, uint8_t *buffer, int max_len)
+int bundle_encode_bundle(struct mmem *bundlemem, uint8_t *buffer, int max_len)
 {
 	uint32_t value;
 	uint8_t offs = 0, blklen_offs, i;
@@ -510,7 +510,7 @@ uint8_t bundle_encode_bundle(struct mmem *bundlemem, uint8_t *buffer, int max_le
 	return offs;
 }
 
-static uint8_t bundle_encode_block(struct bundle_block_t *block, uint8_t *buffer, uint8_t max_len)
+static int bundle_encode_block(struct bundle_block_t *block, uint8_t *buffer, uint8_t max_len)
 {
 	uint8_t offs = 0;
 	int ret;
